Split out gmresult, gmresultplayer, gmregcommon
Attempting to split up our massive gm_1601 file.

Punted on most of the headers, the files are just reusing gm_1601.h
for now. Splitting the header would require changes to many many files
that currently just include gm_1601.h.
