+++ config/GALE01/symbols.txt
@@ -22027,11 +22027,11 @@ lbl_803D6488 = .data:0x803D6488; // type:object size:0xD0
 lbl_803D6558 = .data:0x803D6558; // type:object size:0x300
 lbl_803D6858 = .data:0x803D6858; // type:object size:0x20
 lbl_803D6878 = .data:0x803D6878; // type:object size:0x20 data:byte
-lbl_803D6898 = .data:0x803D6898; // type:object size:0x34 data:string
+lbl_803D6898 = .data:0x803D6898; // type:object size:0x33 data:string
 lbl_803D68CC = .data:0x803D68CC; // type:object size:0x34
 lbl_803D6900 = .data:0x803D6900; // type:object size:0x32 data:string
-lbl_803D6934 = .data:0x803D6934; // type:object size:0xC data:string
-lbl_803D6940 = .data:0x803D6940; // type:object size:0x34 data:string
+lbl_803D6934 = .data:0x803D6934; // type:object size:0xB data:string
+lbl_803D6940 = .data:0x803D6940; // type:object size:0x32 data:string
 lbl_803D6974 = .data:0x803D6974; // type:object size:0x2C data:string
 lbl_803D69A0 = .data:0x803D69A0; // type:object size:0x2C data:string
 lbl_803D69CC = .data:0x803D69CC; // type:object size:0x2C data:string
@@ -25737,19 +25737,19 @@ lbl_804D3F58 = .sdata:0x804D3F58; // type:object size:0x8
 lbl_804D3F60 = .sdata:0x804D3F60; // type:object size:0x8
 lbl_804D3F68 = .sdata:0x804D3F68; // type:object size:0x2 data:string
 lbl_804D3F6C = .sdata:0x804D3F6C; // type:object size:0x2 data:string
-lbl_804D3F70 = .sdata:0x804D3F70; // type:object size:0x8 data:string
-lbl_804D3F78 = .sdata:0x804D3F78; // type:object size:0x8 data:string
-lbl_804D3F80 = .sdata:0x804D3F80; // type:object size:0x8 data:string
-lbl_804D3F88 = .sdata:0x804D3F88; // type:object size:0x4 data:string
+lbl_804D3F70 = .sdata:0x804D3F70; // type:object size:0x6 data:string
+lbl_804D3F78 = .sdata:0x804D3F78; // type:object size:0x7 data:string
+lbl_804D3F80 = .sdata:0x804D3F80; // type:object size:0x7 data:string
+lbl_804D3F88 = .sdata:0x804D3F88; // type:object size:0x3 data:string
 lbl_804D3F8C = .sdata:0x804D3F8C; // type:object size:0x8
-lbl_804D3F94 = .sdata:0x804D3F94; // type:object size:0x4 data:string
-lbl_804D3F98 = .sdata:0x804D3F98; // type:object size:0x8 data:string
+lbl_804D3F94 = .sdata:0x804D3F94; // type:object size:0x3 data:string
+lbl_804D3F98 = .sdata:0x804D3F98; // type:object size:0x5 data:string
 lbl_804D3FA0 = .sdata:0x804D3FA0; // type:object size:0x4
 lbl_804D3FA4 = .sdata:0x804D3FA4; // type:object size:0x4
 lbl_804D3FA8 = .sdata:0x804D3FA8; // type:object size:0x8
 lbl_804D3FB0 = .sdata:0x804D3FB0; // type:object size:0x2 data:string
-lbl_804D3FB4 = .sdata:0x804D3FB4; // type:object size:0x8 data:string
-lbl_804D3FBC = .sdata:0x804D3FBC; // type:object size:0xC data:string
+lbl_804D3FB4 = .sdata:0x804D3FB4; // type:object size:0x7 data:string
+lbl_804D3FBC = .sdata:0x804D3FBC; // type:object size:0x5 data:string
 lbl_804D3FC8 = .sdata:0x804D3FC8; // type:object size:0x8 data:4byte
 lbl_804D3FD0 = .sdata:0x804D3FD0; // type:object size:0x4 data:4byte
 lbl_804D3FD4 = .sdata:0x804D3FD4; // type:object size:0x4 data:4byte
