+++ configure.py
@@ -853,6 +853,12 @@ config.libs = [
             Object(NonMatching, "melee/gm/gmmain_lib.c"),
             Object(Matching, "melee/gm/gmmain.c"),
             Object(NonMatching, "melee/gm/gm_1601.c"),
+            Object(NonMatching, "melee/gm/gmresult.c"),
+            Object(NonMatching, "melee/gm/gmresultplayer.c"),
+            Object(NonMatching, "melee/gm/gm_17AD.c"),
+            Object(NonMatching, "melee/gm/gm_17BA.c"),
+            Object(NonMatching, "melee/gm/gmregcommon.c"),
+            Object(NonMatching, "melee/gm/gm_17C0.c"),
             Object(NonMatching, "melee/gm/gmtou.c"),
             Object(NonMatching, "melee/gm/gm_19EF.c"),
             Object(NonMatching, "melee/gm/gmpause.c"),
