+++ src/melee/gm/gm_1601.h
@@ -1,6 +1,8 @@
 #ifndef GALE01_1601C4
 #define GALE01_1601C4
 
+#include <melee/gm/gm_17C0.h>
+
 #include <platform.h>
 #include <placeholder.h>
 
@@ -537,309 +539,7 @@
 /* 17BE84 */ s32 gm_8017BE84(u32 arg0);
 /* 17BE8C */ UNK_RET gm_8017BE8C(UNK_PARAMS);
 /* 17BEC8 */ UNK_RET gm_8017BEC8(UNK_PARAMS);
-/* 17C0C8 */ UNK_RET fn_8017C0C8(UNK_PARAMS);
-/* 17C1A4 */ UNK_RET fn_8017C1A4(UNK_PARAMS);
-/* 17C71C */ UNK_RET fn_8017C71C(UNK_PARAMS);
-/* 17C7A0 */ UNK_RET fn_8017C7A0(UNK_PARAMS);
-/* 17C7EC */ UNK_RET fn_8017C7EC(UNK_PARAMS);
-/* 17C838 */ UNK_RET gm_8017C838(UNK_PARAMS);
-/* 17C984 */ UNK_RET gm_8017C984(UNK_PARAMS);
-/* 17C9A8 */ UNK_RET gm_8017C9A8(UNK_PARAMS);
-/* 17CA38 */ UNK_RET gm_8017CA38(UNK_PARAMS);
-/* 17CBAC */ UNK_RET gm_8017CBAC(UNK_PARAMS);
-/* 17CD94 */ UNK_RET gm_8017CD94(UNK_PARAMS);
-/* 17CE34 */ UNK_RET gm_8017CE34(UNK_PARAMS);
-/* 17D7AC */ UNK_RET gm_8017D7AC(UNK_PARAMS);
-/* 17D9C0 */ UNK_RET fn_8017D9C0(UNK_PARAMS);
-/* 17DB58 */ void gm_8017DB58(void* arg0);
-/* 17DB6C */ u8 gm_8017DB6C(gm_8017DB6C_arg0_t* arg0, ssize_t index);
-/* 17DB78 */ u8 gm_8017DB78(gm_8017DB6C_arg0_t* arg0, ssize_t index);
-/* 17DB88 */ UNK_RET gm_8017DB88(UNK_PARAMS);
-/* 17DD7C */ UNK_RET fn_8017DD7C(UNK_PARAMS);
-/* 17DE54 */ UNK_RET fn_8017DE54(UNK_PARAMS);
-/* 17DEC8 */ UNK_RET fn_8017DEC8(UNK_PARAMS);
-/* 17DF28 */ UNK_RET fn_8017DF28(UNK_PARAMS);
-/* 17DF90 */ UNK_RET fn_8017DF90(UNK_PARAMS);
-/* 17DFF4 */ UNK_RET gm_8017DFF4(UNK_PARAMS);
-/* 17E068 */ bool gm_8017E068(void);
-/* 17E0E4 */ UNK_RET fn_8017E0E4(UNK_PARAMS);
-/* 17E160 */ UNK_RET fn_8017E160(UNK_PARAMS);
-/* 17E21C */ UNK_RET fn_8017E21C(UNK_PARAMS);
-/* 17E280 */ UNK_T gm_8017E280(u16, bool);
-/* 17E318 */ UNK_RET fn_8017E318(UNK_PARAMS);
-/* 17E3C8 */ UNK_RET fn_8017E3C8(UNK_PARAMS);
-/* 17E424 */ UNK_T gm_8017E424(void);
-/* 17E430 */ u8 gm_8017E430(void);
-/* 17E440 */ u8 gm_8017E440(void);
-/* 17E48C */ UNK_RET gm_8017E48C(UNK_PARAMS);
-/* 17E4C4 */ UNK_RET gm_8017E4C4(UNK_PARAMS);
-/* 17E500 */ UNK_RET gm_8017E500(UNK_PARAMS);
-/* 17E528 */ UNK_RET gm_8017E528(UNK_PARAMS);
-/* 17E578 */ UNK_RET gm_8017E578(UNK_PARAMS);
-/* 17E5C8 */ UNK_RET gm_8017E5C8(UNK_PARAMS);
-/* 17E5FC */ UNK_RET gm_8017E5FC(UNK_PARAMS);
-/* 17E630 */ UNK_RET gm_8017E630(UNK_PARAMS);
-/* 17E664 */ UNK_RET gm_8017E664(UNK_PARAMS);
-/* 17E6B4 */ UNK_RET gm_8017E6B4(UNK_PARAMS);
-/* 17E704 */ UNK_RET gm_8017E704(UNK_PARAMS);
-/* 17E738 */ UNK_RET gm_8017E738(UNK_PARAMS);
-/* 17E76C */ UNK_RET gm_8017E76C(UNK_PARAMS);
-/* 17E7A0 */ UNK_RET gm_8017E7A0(UNK_PARAMS);
-/* 17E7E0 */ UNK_RET gm_8017E7E0(UNK_PARAMS);
-/* 17E7FC */ UNK_RET gm_8017E7FC(UNK_PARAMS);
-/* 17E8A4 */ UNK_RET fn_8017E8A4(UNK_PARAMS);
-/* 17EB30 */ UNK_T gm_8017EB30(void);
-/* 17EB3C */ UNK_RET gm_8017EB3C(UNK_PARAMS);
-/* 17EB64 */ UNK_RET gm_8017EB64(UNK_PARAMS);
-/* 17EB98 */ UNK_RET gm_8017EB98(UNK_PARAMS);
-/* 17EBCC */ UNK_RET gm_8017EBCC(UNK_PARAMS);
-/* 17EC00 */ UNK_RET gm_8017EC00(UNK_PARAMS);
-/* 17EC50 */ UNK_RET gm_8017EC50(UNK_PARAMS);
-/* 17ECA0 */ UNK_RET gm_8017ECA0(UNK_PARAMS);
-/* 17ECD4 */ UNK_RET gm_8017ECD4(UNK_PARAMS);
-/* 17ED08 */ UNK_RET gm_8017ED08(UNK_PARAMS);
-/* 17ED3C */ UNK_RET gm_8017ED3C(UNK_PARAMS);
-/* 17ED8C */ UNK_RET gm_8017ED8C(UNK_PARAMS);
-/* 17EDDC */ UNK_RET fn_8017EDDC(UNK_PARAMS);
-/* 17EE40 */ UNK_RET fn_8017EE40(UNK_PARAMS);
-/* 17F008 */ UNK_RET fn_8017F008(UNK_PARAMS);
-/* 17F09C */ UNK_RET fn_8017F09C(UNK_PARAMS);
-/* 17F14C */ UNK_RET fn_8017F14C(UNK_PARAMS);
-/* 17F1B8 */ UNK_RET fn_8017F1B8(UNK_PARAMS);
-/* 17F2A4 */ UNK_RET fn_8017F2A4(UNK_PARAMS);
-/* 17F47C */ UNK_RET fn_8017F47C(UNK_PARAMS);
-/* 17F608 */ UNK_RET fn_8017F608(UNK_PARAMS);
-/* 17FA1C */ UNK_RET fn_8017FA1C(UNK_PARAMS);
-/* 17FBA4 */ UNK_RET fn_8017FBA4(UNK_PARAMS);
-/* 17FE54 */ UNK_RET fn_8017FE54(UNK_PARAMS);
-/* 17FF1C */ UNK_RET fn_8017FF1C(UNK_PARAMS);
-/* 1803FC */ UNK_RET fn_801803FC(UNK_PARAMS);
-/* 180630 */ UNK_RET fn_80180630(UNK_PARAMS);
-/* 180AC0 */ UNK_RET fn_80180AC0(UNK_PARAMS);
-/* 180AE4 */ s32 gm_80180AE4(void);
-/* 180AF4 */ UNK_RET gm_80180AF4(UNK_PARAMS);
-/* 180B18 */ UNK_RET gm_80180B18(UNK_PARAMS);
-/* 180BA0 */ UNK_RET gm_80180BA0(UNK_PARAMS);
-/* 180C14 */ UNK_RET fn_80180C14(UNK_PARAMS);
-/* 180C60 */ UNK_RET fn_80180C60(UNK_PARAMS);
-/* 181598 */ UNK_RET fn_80181598(UNK_PARAMS);
-/* 181708 */ UNK_RET fn_80181708(UNK_PARAMS);
-/* 181998 */ UNK_RET gm_80181998(UNK_PARAMS);
-/* 181A00 */ void gm_80181A00(s32 arg0, s32 arg1);
-/* 181A24 */ u16 gm_80181A24(void);
-/* 181A34 */ s32 gm_80181A34(void);
-/* 181A44 */ UNK_RET gm_80181A44(UNK_PARAMS);
-/* 181AC8 */ UNK_RET gm_80181AC8(UNK_PARAMS);
-/* 181B64 */ UNK_RET gm_80181B64(UNK_PARAMS);
-/* 181BFC */ UNK_RET fn_80181BFC(UNK_PARAMS);
-/* 181C80 */ UNK_RET fn_80181C80(UNK_PARAMS);
-/* 181E18 */ UNK_RET fn_80181E18(UNK_PARAMS);
-/* 182174 */ UNK_RET gm_80182174(UNK_PARAMS);
-/* 182510 */ UNK_RET gm_80182510(UNK_PARAMS);
-/* 182554 */ UNK_RET gm_80182554(UNK_PARAMS);
-/* 182578 */ UNK_RET gm_80182578(UNK_PARAMS);
-/* 182B5C */ UNK_RET fn_80182B5C(UNK_PARAMS);
-/* 182DF0 */ UNK_RET gm_80182DF0(UNK_PARAMS);
-/* 182F40 */ UNK_RET fn_80182F40(UNK_PARAMS);
-/* 183218 */ UNK_RET gm_80183218(UNK_PARAMS);
-/* 18325C */ UNK_RET fn_8018325C(UNK_PARAMS);
-/* 184138 */ UNK_RET fn_80184138(UNK_PARAMS);
-/* 1849E0 */ UNK_RET fn_801849E0(UNK_PARAMS);
-/* 184A04 */ UNK_RET fn_80184A04(UNK_PARAMS);
-/* 184A28 */ UNK_RET fn_80184A28(UNK_PARAMS);
-/* 184A4C */ UNK_RET fn_80184A4C(UNK_PARAMS);
-/* 184A70 */ UNK_RET fn_80184A70(UNK_PARAMS);
-/* 184A94 */ UNK_RET fn_80184A94(UNK_PARAMS);
-/* 184AB8 */ UNK_RET fn_80184AB8(UNK_PARAMS);
-/* 18504C */ UNK_RET fn_8018504C(UNK_PARAMS);
-/* 1851C0 */ UNK_RET fn_801851C0(UNK_PARAMS);
-/* 1852FC */ UNK_RET fn_801852FC(UNK_PARAMS);
-/* 185408 */ UNK_RET fn_80185408(UNK_PARAMS);
-/* 1855BC */ UNK_RET fn_801855BC(UNK_PARAMS);
-/* 18564C */ UNK_RET fn_8018564C(UNK_PARAMS);
-/* 18569C */ UNK_RET fn_8018569C(UNK_PARAMS);
-/* 18575C */ UNK_RET fn_8018575C(UNK_PARAMS);
-/* 1857C4 */ UNK_RET fn_801857C4(UNK_PARAMS);
-/* 1859C8 */ UNK_RET fn_801859C8(UNK_PARAMS);
-/* 185A0C */ UNK_RET fn_80185A0C(UNK_PARAMS);
-/* 185D64 */ UNK_RET fn_80185D64(UNK_PARAMS);
-/* 185E34 */ UNK_RET fn_80185E34(UNK_PARAMS);
-/* 185F5C */ UNK_RET fn_80185F5C(UNK_PARAMS);
-/* 186080 */ UNK_RET fn_80186080(UNK_PARAMS);
-/* 1861B8 */ UNK_RET fn_801861B8(UNK_PARAMS);
-/* 186400 */ UNK_RET fn_80186400(UNK_PARAMS);
-/* 186634 */ UNK_RET fn_80186634(UNK_PARAMS);
-/* 186DFC */ UNK_RET gm_80186DFC_OnFrame(UNK_PARAMS);
-/* 186E30 */ void gm_80186E30_OnEnter(UNK_T);
-/* 186EFC */ UNK_RET fn_80186EFC(UNK_PARAMS);
-/* 186F6C */ UNK_RET fn_80186F6C(UNK_PARAMS);
-/* 1873F0 */ UNK_RET fn_801873F0(UNK_PARAMS);
-/* 187494 */ UNK_RET fn_80187494(UNK_PARAMS);
-/* 1874FC */ UNK_RET fn_801874FC(UNK_PARAMS);
-/* 187714 */ UNK_RET fn_80187714(UNK_PARAMS);
-/* 18776C */ UNK_RET gm_8018776C_OnFrame(UNK_PARAMS);
-/* 1877A8 */ void gm_801877A8_OnEnter(UNK_T);
-/* 187910 */ UNK_RET fn_80187910(UNK_PARAMS);
-/* 187AB4 */ UNK_RET fn_80187AB4(UNK_PARAMS);
-/* 187C9C */ UNK_RET fn_80187C9C(UNK_PARAMS);
-/* 187CF4 */ UNK_RET fn_80187CF4(UNK_PARAMS);
-/* 187F48 */ void gm_80187F48_OnEnter(UNK_T);
-/* 188364 */ void gm_80188364_OnLeave(UNK_T);
-/* 18838C */ UNK_RET gm_8018838C_OnFrame(UNK_PARAMS);
-/* 1883C0 */ UNK_RET gm_801883C0(UNK_PARAMS);
-/* 18841C */ bool gm_8018841C(void);
-/* 188454 */ int gm_80188454(int idx);
-/* 18846C */ UNK_RET fn_8018846C(UNK_PARAMS);
-/* 1884F8 */ UNK_RET fn_801884F8(UNK_PARAMS);
-/* 188550 */ UNK_RET fn_80188550(UNK_PARAMS);
-/* 188644 */ UNK_RET fn_80188644(UNK_PARAMS);
-/* 188738 */ UNK_RET fn_80188738(UNK_PARAMS);
-/* 188910 */ UNK_RET fn_80188910(UNK_PARAMS);
-/* 188B3C */ UNK_RET fn_80188B3C(UNK_PARAMS);
-/* 188D3C */ UNK_RET fn_80188D3C(UNK_PARAMS);
-/* 188EE8 */ UNK_RET fn_80188EE8(UNK_PARAMS);
-/* 1891F4 */ UNK_RET fn_801891F4(UNK_PARAMS);
-/* 189B88 */ UNK_RET fn_80189B88(UNK_PARAMS);
-/* 189CDC */ UNK_RET gm_80189CDC(UNK_PARAMS);
-/* 18A000 */ UNK_RET fn_8018A000(UNK_PARAMS);
-/* 18A160 */ UNK_RET gm_8018A160(UNK_PARAMS);
-/* 18A188 */ UNK_RET gm_8018A188(UNK_PARAMS);
-/* 18A1D8 */ UNK_RET gm_8018A1D8(UNK_PARAMS);
-/* 18A228 */ UNK_RET gm_8018A228(UNK_PARAMS);
-/* 18A25C */ UNK_RET gm_8018A25C(UNK_PARAMS);
-/* 18A290 */ UNK_RET gm_8018A290(UNK_PARAMS);
-/* 18A2C4 */ UNK_RET gm_8018A2C4(UNK_PARAMS);
-/* 18A314 */ UNK_RET gm_8018A314(UNK_PARAMS);
-/* 18A364 */ UNK_RET fn_8018A364(UNK_PARAMS);
-/* 18A514 */ UNK_RET fn_8018A514(UNK_PARAMS);
-/* 18A970 */ UNK_RET fn_8018A970(UNK_PARAMS);
-/* 18AA74 */ UNK_RET fn_8018AA74(UNK_PARAMS);
-/* 18B090 */ UNK_RET fn_8018B090(UNK_PARAMS);
-/* 18C8D4 */ UNK_RET fn_8018C8D4(UNK_PARAMS);
-/* 18D50C */ UNK_RET fn_8018D50C(UNK_PARAMS);
-/* 18DC18 */ UNK_RET fn_8018DC18(UNK_PARAMS);
-/* 18DF68 */ UNK_RET fn_8018DF68(UNK_PARAMS);
-/* 18E46C */ UNK_RET fn_8018E46C(UNK_PARAMS);
-/* 18E618 */ UNK_RET fn_8018E618(UNK_PARAMS);
-/* 18E85C */ UNK_RET fn_8018E85C(UNK_PARAMS);
-/* 18EC48 */ UNK_RET fn_8018EC48(UNK_PARAMS);
-/* 18EC7C */ UNK_RET fn_8018EC7C(UNK_PARAMS);
-/* 18ECA8 */ UNK_RET fn_8018ECA8(UNK_PARAMS);
-/* 18F00C */ UNK_RET fn_8018F00C(UNK_PARAMS);
-/* 18F1B0 */ UNK_RET gm_8018F1B0(UNK_PARAMS);
-/* 18F310 */ UNK_RET fn_8018F310(UNK_PARAMS);
-// /* 18F3BC */ UNK_RET fn_8018F3BC(UNK_PARAMS);
-/* 18F3D0 */ UNK_RET fn_8018F3D0(UNK_PARAMS);
-/* 18F410 */ UNK_RET fn_8018F410(UNK_PARAMS);
-/* 18F4A0 */ UNK_RET fn_8018F4A0(UNK_PARAMS);
-/* 18F508 */ UNK_RET fn_8018F508(UNK_PARAMS);
-/* 18F5F0 */ UNK_RET fn_8018F5F0(UNK_PARAMS);
-/* 18F62C */ s32 fn_8018F62C(UNK_T arg0);
-/* 18F634 */ UNK_T gm_8018F634(void);
-/* 18F640 */ UNK_RET fn_8018F640(UNK_PARAMS);
-/* 18F674 */ UNK_RET fn_8018F674(UNK_PARAMS);
-/* 18F6A8 */ UNK_RET fn_8018F6A8(UNK_PARAMS);
-/* 18F6DC */ UNK_RET fn_8018F6DC(UNK_PARAMS);
-/* 18F6FC */ UNK_RET fn_8018F6FC(UNK_PARAMS);
-/* 18F71C */ UNK_RET fn_8018F71C(UNK_PARAMS);
-/* 18F74C */ UNK_RET fn_8018F74C(UNK_PARAMS);
-/* 18F808 */ UNK_RET fn_8018F808(UNK_PARAMS);
-/* 18F888 */ UNK_RET fn_8018F888(UNK_PARAMS);
-/* 18FA24 */ UNK_RET fn_8018FA24(UNK_PARAMS);
-/* 18FBD8 */ void fn_8018FBD8(UNK_T arg0, s32 arg1);
-/* 18FBE0 */ UNK_RET fn_8018FBE0(UNK_PARAMS);
-/* 18FDC4 */ UNK_RET fn_8018FDC4(UNK_PARAMS);
-/* 18FF9C */ UNK_RET fn_8018FF9C(UNK_PARAMS);
-/* 190174 */ UNK_RET fn_80190174(UNK_PARAMS);
-/* 1901F8 */ UNK_RET fn_801901F8(UNK_PARAMS);
-/* 19027C */ UNK_RET fn_8019027C(UNK_PARAMS);
-/* 1902F0 */ UNK_RET fn_801902F0(UNK_PARAMS);
-/* 19035C */ UNK_RET fn_8019035C(UNK_PARAMS);
-/* 19044C */ UNK_RET fn_8019044C(UNK_PARAMS);
-/* 190480 */ UNK_RET fn_80190480(UNK_PARAMS);
-/* 1904D0 */ UNK_RET fn_801904D0(UNK_PARAMS);
-/* 190520 */ UNK_RET fn_80190520(UNK_PARAMS);
-/* 1905F0 */ UNK_RET gm_801905F0(UNK_PARAMS);
-/* 190ABC */ UNK_RET fn_80190ABC(UNK_PARAMS);
-/* 190EA4 */ UNK_RET gm_80190EA4(UNK_PARAMS);
-/* 190FE4 */ UNK_RET gm_80190FE4(UNK_PARAMS);
-/* 1910E0 */ UNK_RET fn_801910E0(UNK_PARAMS);
-/* 191154 */ UNK_RET fn_80191154(UNK_PARAMS);
-/* 191240 */ UNK_RET fn_80191240(UNK_PARAMS);
-/* 1913BC */ UNK_RET fn_801913BC(UNK_PARAMS);
-/* 191678 */ UNK_RET fn_80191678(UNK_PARAMS);
-/* 19175C */ UNK_RET fn_8019175C(UNK_PARAMS);
-/* 1918F0 */ UNK_RET fn_801918F0(UNK_PARAMS);
-/* 191A54 */ UNK_RET fn_80191A54(UNK_PARAMS);
-/* 191B5C */ UNK_RET fn_80191B5C(UNK_PARAMS);
-/* 191CA4 */ UNK_RET fn_80191CA4(UNK_PARAMS);
-/* 191D38 */ UNK_RET fn_80191D38(UNK_PARAMS);
-/* 191E9C */ UNK_RET fn_80191E9C(UNK_PARAMS);
-/* 191FD4 */ UNK_RET fn_80191FD4(UNK_PARAMS);
-/* 19237C */ UNK_RET fn_8019237C(UNK_PARAMS);
-/* 19249C */ UNK_RET fn_8019249C(UNK_PARAMS);
-/* 192690 */ UNK_RET fn_80192690(UNK_PARAMS);
-/* 192758 */ UNK_RET fn_80192758(UNK_PARAMS);
-/* 192938 */ UNK_RET fn_80192938(UNK_PARAMS);
-/* 192BB0 */ UNK_RET fn_80192BB0(UNK_PARAMS);
-/* 192E6C */ UNK_RET fn_80192E6C(UNK_PARAMS);
-/* 193230 */ UNK_RET fn_80193230(UNK_PARAMS);
-/* 193308 */ UNK_RET fn_80193308(UNK_PARAMS);
-/* 1935B8 */ UNK_RET fn_801935B8(UNK_PARAMS);
-/* 1937C4 */ UNK_RET fn_801937C4(UNK_PARAMS);
-/* 193B58 */ UNK_RET fn_80193B58(UNK_PARAMS);
-/* 193FCC */ UNK_RET fn_80193FCC(UNK_PARAMS);
-/* 194658 */ UNK_RET fn_80194658(UNK_PARAMS);
-/* 1949B4 */ UNK_RET fn_801949B4(UNK_PARAMS);
-/* 194BC4 */ UNK_RET fn_80194BC4(UNK_PARAMS);
-/* 194D84 */ UNK_RET fn_80194D84(UNK_PARAMS);
-/* 194F30 */ UNK_RET fn_80194F30(UNK_PARAMS);
-/* 1953C8 */ UNK_RET fn_801953C8(UNK_PARAMS);
-/* 195AF0 */ UNK_RET fn_80195AF0(UNK_PARAMS);
-/* 195CCC */ UNK_RET fn_80195CCC(UNK_PARAMS);
-/* 19610C */ UNK_RET fn_8019610C(UNK_PARAMS);
-/* 19628C */ void gm_8019628C_OnFrame(void);
-/* 1963B4 */ void gm_801963B4_OnEnter(UNK_T);
-/* 1964A4 */ void gm_801964A4_OnLeave(UNK_T);
-/* 196510 */ UNK_RET fn_80196510(UNK_PARAMS);
-/* 19655C */ s32 fn_8019655C(void);
-/* 196564 */ UNK_RET fn_80196564(UNK_PARAMS);
-/* 196594 */ UNK_RET fn_80196594(UNK_PARAMS);
-/* 1965C4 */ UNK_RET fn_801965C4(UNK_PARAMS);
-/* 196684 */ UNK_RET fn_80196684(UNK_PARAMS);
-/* 1967E0 */ UNK_RET fn_801967E0(UNK_PARAMS);
-/* 196CF8 */ UNK_RET fn_80196CF8(UNK_PARAMS);
-/* 196DBC */ UNK_RET fn_80196DBC(UNK_PARAMS);
-/* 196E30 */ UNK_RET fn_80196E30(UNK_PARAMS);
-/* 196EEC */ UNK_RET fn_80196EEC(UNK_PARAMS);
-/* 196FFC */ UNK_RET fn_80196FFC(UNK_PARAMS);
-/* 1973F8 */ UNK_RET fn_801973F8(UNK_PARAMS);
-/* 1975C8 */ UNK_RET fn_801975C8(UNK_PARAMS);
-/* 1976D4 */ UNK_RET fn_801976D4(UNK_PARAMS);
-/* 1977AC */ UNK_RET fn_801977AC(UNK_PARAMS);
-/* 197AF0 */ UNK_RET fn_80197AF0(UNK_PARAMS);
-/* 197D4C */ UNK_RET fn_80197D4C(UNK_PARAMS);
-/* 197E18 */ UNK_RET fn_80197E18(UNK_PARAMS);
-/* 197FD8 */ UNK_RET fn_80197FD8(UNK_PARAMS);
-/* 1981A0 */ UNK_RET fn_801981A0(UNK_PARAMS);
-/* 1983E4 */ UNK_RET fn_801983E4(UNK_PARAMS);
-/* 198584 */ UNK_RET fn_80198584(UNK_PARAMS);
-/* 1985D4 */ UNK_RET fn_801985D4(UNK_PARAMS);
-/* 198824 */ UNK_RET fn_80198824(UNK_PARAMS);
-/* 198BA0 */ UNK_RET fn_80198BA0(UNK_PARAMS);
-/* 198C60 */ UNK_RET fn_80198C60(UNK_PARAMS);
-/* 198D18 */ UNK_RET fn_80198D18(UNK_PARAMS);
-/* 198EBC */ UNK_RET fn_80198EBC(UNK_PARAMS);
-/* 199AF0 */ UNK_RET fn_80199AF0(UNK_PARAMS);
-/* 19A158 */ UNK_RET fn_8019A158(UNK_PARAMS);
-/* 19A71C */ UNK_RET fn_8019A71C(UNK_PARAMS);
-/* 19A828 */ UNK_RET gm_8019A828(UNK_PARAMS);
-/* 19A86C */ UNK_RET fn_8019A86C(UNK_PARAMS);
-/* 19AF50 */ UNK_RET fn_8019AF50(UNK_PARAMS);
-/* 19B2DC */ UNK_RET gm_8019B2DC_OnFrame(UNK_PARAMS);
-/* 19B458 */ UNK_RET fn_8019B458(UNK_PARAMS);
-/* 19B81C */ UNK_RET fn_8019B81C(UNK_PARAMS);
-/* 19B860 */ UNK_RET fn_8019B860(UNK_PARAMS);
+
 /* 473A18 */ extern UNK_T gm_80473A18;
 
 #endif
