+++ src/melee/gm/gm_17AD.c
@@ -1,0 +1,72 @@
+#include "gm_1601.h"
+
+#include "gm_1601.static.h"
+
+
+/// #fn_8017AD04
+
+/// #fn_8017AD28
+
+/// #fn_8017AD48
+
+/// #fn_8017AD78
+
+/// #fn_8017ADA8
+
+/// #fn_8017AE0C
+
+/// #fn_8017AE70
+
+/// #fn_8017AED8
+
+/// #fn_8017AF40
+
+/// #fn_8017AFA8
+
+/// #fn_8017B010
+
+/// #fn_8017B07C
+
+/// #fn_8017B0E4
+
+/// #fn_8017B14C
+
+/// #fn_8017B1B4
+
+/// #fn_8017B21C
+
+/// #fn_8017B280
+
+/// #fn_8017B2E4
+
+/// #fn_8017B348
+
+/// #fn_8017B3AC
+
+/// #fn_8017B410
+
+/// #fn_8017B4D0
+
+/// #fn_8017B534
+
+/// #fn_8017B598
+
+/// #fn_8017B5FC
+
+/// #fn_8017B660
+
+/// #fn_8017B6C4
+
+/// #fn_8017B728
+
+/// #fn_8017B78C
+
+/// #fn_8017B7F0
+
+/// #fn_8017B854
+
+/// #fn_8017B8B8
+
+/// #fn_8017B91C
+
+/// #fn_8017B9F4
