+++ src/melee/gm/gm_17BA.c
@@ -1,0 +1,16 @@
+#include "gm_1601.h"
+
+#include "gm_1601.static.h"
+
+
+/// #fn_8017BACC
+
+/// #fn_8017BB30
+
+/// #fn_8017BB94
+
+/// #fn_8017BC50
+
+/// #fn_8017BD0C
+
+/// #fn_8017BDC8
