+++ src/melee/gm/gmregcommon.c
@@ -1,0 +1,13 @@
+#include "gm_1601.h"
+
+#include "gm_1601.static.h"
+
+
+s32 gm_8017BE84(u32 arg0)
+{
+    return (arg0 >> 3) & 0x1F;
+}
+
+/// #gm_8017BE8C
+
+/// #gm_8017BEC8
