+++ src/melee/gm/gmresult.c
@@ -1,0 +1,73 @@
+#include "gm_1601.h"
+
+#include "gm_1601.static.h"
+
+
+int fn_80174274(void)
+{
+    return lbl_8046DBE8.x94;
+}
+
+/// #fn_80174284
+
+/// #fn_80174338
+
+/// #fn_8017435C
+
+/// #fn_80174380
+
+/// #gm_801743A4
+
+/// #fn_801743C4
+
+/// #fn_80174468
+
+/// #fn_801748EC
+
+/// #fn_80174920
+
+/// #fn_801749B8
+
+/// #fn_80174A60
+
+/// #fn_80174B4C
+
+/// #fn_80174FD0
+
+/// #fn_80175038
+
+/// #fn_8017507C
+
+/// #fn_80175240
+
+/// #fn_8017556C
+
+/// #fn_801756E0
+
+/// #fn_80175880
+
+/// #fn_80175A94
+
+/// #fn_80175C5C
+
+/// #fn_80175D34
+
+/// #fn_80175DC8
+
+/// #fn_80176A6C
+
+/// #fn_80176BCC
+
+/// #fn_80176BF0
+
+/// #fn_80176D18
+
+/// #fn_80176D3C
+
+/// #fn_80176F60
+
+/// #fn_801771C0
+
+/// #gm_80177368_OnEnter
+
+/// #gm_80177704_OnLeave
