+++ src/melee/gm/gmresultplayer.c
@@ -1,0 +1,67 @@
+#include "gm_1601.h"
+
+#include "gm_1601.static.h"
+
+
+/// #gm_80177724
+
+/// #fn_80177748
+
+/// #fn_80177920
+
+/// #fn_80177B7C
+
+/// #fn_80177DD0
+
+/// #fn_80178050
+
+/// #fn_801785B0
+
+/// #fn_80178BB4
+
+/// #fn_801791E4
+
+/// #fn_80179350
+
+/// #fn_801795D4
+
+/// #fn_801796F0
+
+/// #fn_80179854
+
+/// #fn_80179990
+
+/// #fn_80179D3C
+
+/// #fn_80179D60
+
+/// #fn_80179D84
+
+/// #fn_80179DA8
+
+/// #fn_80179DCC
+
+/// #fn_80179E34
+
+/// #fn_80179E9C
+
+/// #fn_80179F04
+
+void fn_80179F6C(int idx, int value)
+{
+    lbl_8046E38C[idx] = value;
+}
+
+/// #fn_80179F84
+
+/// #fn_8017A004
+
+/// #fn_8017A078
+
+/// #fn_8017A318
+
+/// #fn_8017A67C
+
+/// #fn_8017A9B4
+
+/// #fn_8017AA78
